+++ Tests/src/PublicAPITests/ComponentTests/VideoPlayerComponentTests.cpp
@@ -87,7 +87,7 @@ CSP_PUBLIC_TEST(CSPEngine, VideoTests, VideoPlayerComponentTest)
     EXPECT_EQ(VideoComponent->GetIsAutoResize(), false);
     EXPECT_EQ(VideoComponent->GetCurrentPlayheadPosition(), 0.0f);
     EXPECT_EQ(VideoComponent->GetVideoPlayerSourceType(), VideoPlayerSourceType::AssetSource);
-    EXPECT_EQ(VideoComponent->GetStereoVideoType(), EStereoVideoType::None);
+    EXPECT_EQ(VideoComponent->GetStereoVideoType(), StereoVideoType::None);
     EXPECT_EQ(VideoComponent->GetIsVisible(), true);
     EXPECT_EQ(VideoComponent->GetIsARVisible(), true);
     EXPECT_EQ(VideoComponent->GetIsVirtualVisible(), true);
@@ -113,7 +113,7 @@ CSP_PUBLIC_TEST(CSPEngine, VideoTests, VideoPlayerComponentTest)
     VideoComponent->SetIsAutoResize(true);
     VideoComponent->SetCurrentPlayheadPosition(1.0f);
     VideoComponent->SetVideoPlayerSourceType(VideoPlayerSourceType::URLSource);
-    VideoComponent->SetStereoVideoType(EStereoVideoType::SideBySide);
+    VideoComponent->SetStereoVideoType(StereoVideoType::SideBySide);
     VideoComponent->SetIsVisible(false);
     VideoComponent->SetIsARVisible(false);
     VideoComponent->SetIsVirtualVisible(false);
@@ -133,7 +133,7 @@ CSP_PUBLIC_TEST(CSPEngine, VideoTests, VideoPlayerComponentTest)
     EXPECT_EQ(VideoComponent->GetIsAutoResize(), true);
     EXPECT_EQ(VideoComponent->GetCurrentPlayheadPosition(), 1.0f);
     EXPECT_EQ(VideoComponent->GetVideoPlayerSourceType(), VideoPlayerSourceType::URLSource);
-    EXPECT_EQ(VideoComponent->GetStereoVideoType(), EStereoVideoType::SideBySide);
+    EXPECT_EQ(VideoComponent->GetStereoVideoType(), StereoVideoType::SideBySide);
     EXPECT_EQ(VideoComponent->GetIsVisible(), false);
     EXPECT_EQ(VideoComponent->GetIsARVisible(), false);
     EXPECT_EQ(VideoComponent->GetIsVirtualVisible(), false);
@@ -201,7 +201,7 @@ CSP_PUBLIC_TEST(CSPEngine, VideoTests, VideoPlayerScriptInterfaceTest)
     EXPECT_EQ(VideoPlayerComponent->GetCurrentPlayheadPosition(), 0.0f);
     EXPECT_EQ(VideoPlayerComponent->GetTimeSincePlay(), 0.f);
     EXPECT_EQ(VideoPlayerComponent->GetVideoPlayerSourceType(), VideoPlayerSourceType::AssetSource);
-    EXPECT_EQ(VideoPlayerComponent->GetStereoVideoType(), EStereoVideoType::None);
+    EXPECT_EQ(VideoPlayerComponent->GetStereoVideoType(), StereoVideoType::None);
     EXPECT_EQ(VideoPlayerComponent->GetIsVisible(), true);
     EXPECT_EQ(VideoPlayerComponent->GetIsARVisible(), true);
     EXPECT_EQ(VideoPlayerComponent->GetIsVirtualVisible(), true);
@@ -256,7 +256,7 @@ CSP_PUBLIC_TEST(CSPEngine, VideoTests, VideoPlayerScriptInterfaceTest)
     EXPECT_EQ(VideoPlayerComponent->GetCurrentPlayheadPosition(), 1.0f);
     EXPECT_EQ(VideoPlayerComponent->GetTimeSincePlay(), 1.f);
     EXPECT_EQ(VideoPlayerComponent->GetVideoPlayerSourceType(), VideoPlayerSourceType::URLSource);
-    EXPECT_EQ(VideoPlayerComponent->GetStereoVideoType(), EStereoVideoType::SideBySide);
+    EXPECT_EQ(VideoPlayerComponent->GetStereoVideoType(), StereoVideoType::SideBySide);
     EXPECT_EQ(VideoPlayerComponent->GetIsStereoFlipped(), true);
     EXPECT_EQ(VideoPlayerComponent->GetIsVisible(), false);
     EXPECT_EQ(VideoPlayerComponent->GetIsARVisible(), false);